Fix indexing error for proportional palettes
Palette width wasn't clamped so there was a possibility of off-by-one
when working with very small percentages in proportional palettes. This
became apparent when they were sorted and for large sizes.

Add flag for working with transparent images. First steps of working
with handling k-means for transparent pngs. The flag is hidden and
only good for finding the colors or producing a palette. It's a special
case that will require more logic for outputting a full image. This
will be completed with hopefully a change to generics.