Implement Hamerly's algorithm for Lab and Srgb
Reformat README.md.
Add logic for adaptive switching between Hamerly and naive in the
binary. k=6 seems to be the crossover point for colors when it's
advantageous to use Hamerly over naive. This is not the case in
the lanterns example: there are many similar colors, so distance
calculations end up being performed and the overhead of the cached
distance results likely slows down calculation. The flowers example
has clear winnings probably attributable to clearer segmentation
of colors. Seconds can come off the calculation because of early
convergence in the best case images.

Implementation note:
Square roots were necessary for the bounds calculation to result in
the same calculation of colors as the naive version. The algorithms
should be deterministic and result in the same colors. Tests should
be added to verify this stays the case automatically.