Update dependencies
After submitting the patch upstream, we can remove the `png` dependency
since `save_with_quality` is now available for png to use NoFilter and
Best compression. When adaptive filtering is eventually added, that will
be selected for the filter type.

The decoder should also be faster for png due to the switch to miniz_oxide.